fixed code feedbacks
