+++ .gitignore
@@ -5,3 +5,4 @@ node_modules
 .env
 .vscode
 yarn.lock
+package-lock.json
